feat: query, mutation code 수정
