+++ packages/api/src/index.ts
@@ -1,2 +1,6 @@
-export * from './config/api-instance';
-export * from './config/types';
+export { PetApi } from '../src/api/pet';
+export { StoreApi } from '../src/api/store';
+export { UserApi } from '../src/api/user';
+
+export { apiInstance } from '../src/config/api-instance';
+export { type APIResponseData } from '../src/config/types';
