+++ packages/api/templates/swagger-api-template/api.ejs
@@ -10,7 +10,7 @@ const dataContracts = _.map(modelTypes, "name");
 
 import { HttpClient, RequestParams, ContentType, HttpResponse } from "./<%~ config.fileNames.httpClient %>";
 <% if (dataContracts.length) { %>
-import type { <% ~dataContracts.join(", ") %> } from <%~`"../types/${apiClassName.toLowerCase()}.ts"` %>;
+import type { <% ~dataContracts.join(", ") %> } from <%~`"../types"` %>;
 <% } %>
 import { apiInstance } from '../config/api-instance';
