+++ packages/api/templates/swagger-mutation-template/api.ejs
@@ -10,10 +10,10 @@ const dataContracts = _.map(modelTypes, "name");
 
 import { HttpClient, RequestParams, ContentType, HttpResponse } from "./<%~ config.fileNames.httpClient %>";
 <% if (dataContracts.length) { %>
-import type { <% ~dataContracts.join(", ") %> } from <%~`"../types/swagger-${apiClassName.toLowerCase()}.ts"` %>;
+import type { <% ~dataContracts.join(", ") %> } from <%~`"../types"` %>;
 <% } %>
-import { useMutation } from '@tanstack/vue-query';
-import { <% ~apiClassName %>Api } from '@common/api';
+import { useMutation } from '@tanstack/react-query';
+import { <% ~apiClassName %>Api } from '../';
 
 
 <% for (const route of routes) { %>
