+++ packages/api/templates/swagger-query-template/api.ejs
@@ -10,10 +10,10 @@ const dataContracts = _.map(modelTypes, "name");
 
 import { HttpClient, RequestParams, ContentType, HttpResponse } from "./<%~ config.fileNames.httpClient %>";
 <% if (dataContracts.length) { %>
-import type { <% ~dataContracts.join(", ") %> } from <%~`"../types/swagger-${apiClassName.toLowerCase()}.ts"` %>;
+import type { <% ~dataContracts.join(", ") %> } from <%~`"../types"` %>;
 <% } %>
-import { useQuery } from '@tanstack/vue-query';
-import { APIResponseData, <%~ apiClassName %>Api } from '@common/api';
+import { useQuery } from '@tanstack/react-query';
+import { APIResponseData, <%~ apiClassName %>Api } from '../';
 
 
 <% for (const route of routes) { %>
